move auth/main to views dir
